GUI/requirements.txt
fastapi==0.116.1
numpy==2.2.6
opencv-python-headless==4.12.0.88
tensorflow_cpu==2.20.0
uvicorn[standard]==0.35.0